% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{shen268_cerebellar}
\alias{shen268_cerebellar}
\title{Shen 268 Cerebellar Parcellation}
\usage{
shen268_cerebellar()
}
\value{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (subcortical).
}
\description{
Cerebellar regions from the Shen 268-node functional parcellation
(Shen et al., 2013). Contains 2D polygon geometry and 3D meshes.
}
\examples{
\dontrun{
shen268_cerebellar()
}
}
\references{
Shen X et al. (2013). Groupwise whole-brain parcellation
from resting-state fMRI data for network node identification.
\doi{10.1016/j.neuroimage.2013.05.081}
}
\seealso{
Other ggseg_atlases: 
\code{\link{shen268_cortical}()},
\code{\link{shen268_subcortical}()}
}
\concept{cerebellar_atlases}
\concept{ggseg_atlases}
