% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shen268_cortical}
\alias{shen268_cortical}
\title{Shen 268 Cortical Parcellation}
\format{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (cortical).
}
\usage{
data(shen268_cortical)
}
\description{
Cortical regions from the Shen 268-node functional parcellation
(Shen et al., 2013). Contains 2D polygon geometry for
\code{\link[ggseg:ggbrain]{ggseg::geom_brain()}} and 3D vertex indices for \code{\link[ggseg3d:ggseg3d]{ggseg3d::ggseg3d()}}.
}
\examples{
data(shen268_cortical)
shen268_cortical
}
\references{
Shen X et al. (2013). Groupwise whole-brain parcellation
from resting-state fMRI data for network node identification.
\doi{10.1016/j.neuroimage.2013.05.081}
}
\seealso{
Other ggseg_atlases: 
\code{\link{shen268_subcortical}}
}
\concept{ggseg_atlases}
\keyword{datasets}
