% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shen268_subcortical}
\alias{shen268_subcortical}
\title{Shen 268 Subcortical Parcellation}
\format{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (subcortical).
}
\usage{
data(shen268_subcortical)
}
\description{
Subcortical regions from the Shen 268-node functional parcellation
(Shen et al., 2013). Contains 2D polygon geometry and 3D meshes.
}
\examples{
data(shen268_subcortical)
shen268_subcortical
}
\references{
Shen X et al. (2013). Groupwise whole-brain parcellation
from resting-state fMRI data for network node identification.
\doi{10.1016/j.neuroimage.2013.05.081}
}
\seealso{
Other ggseg_atlases: 
\code{\link{shen268_cortical}}
}
\concept{ggseg_atlases}
\keyword{datasets}
